animate edges when hovering over node while in linking
previously, when creating a edge, the halo highlight effect around a
node which appears when a mouse hovers over the node was “expanded out”
of the node. i.e., it was animated. however, the arrow adjustment was
not.

now it is!

required manually interpolating the values for the bounding edges of
the node rect with d3’s help